New folder/requirements.txt
Flask
bcrypt
python-dotenv
twilio